+++ utilities/user_utilities.py
@@ -238,7 +238,7 @@ def find_user_all_object_events(eventId):
 def delete_user_event_in_building_block(objectId_list):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + current_app.config['AUTHENTICATION_TOKEN']
+        'Authorization': 'Bearer ' + session["id_token"]
     }
     delete_success_list = []
     fail_count = 0
@@ -317,7 +317,7 @@ def approve_user_event(objectId):
 def publish_user_event(eventId):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + current_app.config['AUTHENTICATION_TOKEN']
+        'Authorization': 'Bearer ' + session["id_token"]
     }
 
     try:
@@ -396,7 +396,7 @@ def publish_user_event(eventId):
 def put_user_event(eventId):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + current_app.config['AUTHENTICATION_TOKEN']
+        'Authorization': 'Bearer ' + session["id_token"]
     }
     try:
         # Put event in object, but exclude ID and status
@@ -557,7 +557,7 @@ def populate_event_from_form(post_form, email):
 
     location = post_form.get('location')
     if location != '':
-        new_event['location'] = get_location_details(location)
+        new_event['location'] = get_location_details(location, new_event.get('isVirtual'))
     else:
         new_event['location'] = None
 
@@ -566,8 +566,12 @@ def populate_event_from_form(post_form, email):
     return new_event
 
 
-def get_location_details(location_description):
+def get_location_details(location_description, is_virtual_event):
     location_obj = dict()
+    if is_virtual_event:
+        location_obj['description'] = location_description
+        return location_obj
+
     for excluded_location in Config.EXCLUDED_LOCATION:
         if excluded_location.lower() in location_description.lower():
             location_obj['description'] = location_description