+++ CHANGELOG.md
@@ -16,14 +16,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Event filtering by groups on event listing page [#676](https://github.com/rokwire/events-manager/issues/676)
 
 ### Fixed
+- End dates can't be deleted. [#691](https://github.com/rokwire/events-manager/issues/690)
+- All-day events not working. [#690](https://github.com/rokwire/events-manager/issues/691)
 - User doesn't belong to any group will trigger a redirect loop upon login. [#695](https://github.com/rokwire/events-manager/issues/695)
 - delete events from events manager in local if the events are not in events building blocks.[#684](https://github.com/rokwire/events-manager/issues/684)
 - get_admin_group_ids() will only return first group ID.[#699](https://github.com/rokwire/events-manager/issues/699)
+- Fix failure to publish a free user event. [#697](https://github.com/rokwire/events-manager/issues/697)
+
 ### Changed
 - Crop event title at a fixed number of characters. [#667](https://github.com/rokwire/events-manager/issues/667)
 - Move dates under a second row below the title and move badges to the right of the title. [#666](https://github.com/rokwire/events-manager/issues/666)
 - Split datetime-local control to date and time to fix compatibility issue with firefox and safari. [#479](https://github.com/rokwire/events-manager/issues/479)
-
+- Move Free Event above Cost Description. [#661](https://github.com/rokwire/events-manager/issues/661)
 ## [2.4.1] - 2021-06-23
 ### Added